+++ core/src/main/java/eu/unicore/ucc/actions/OpenTunnel.java
@@ -46,6 +46,7 @@ public class OpenTunnel extends ActionBase {
 				.required(false)
 				.build());
 	}
+
 	@Override
 	public String getName(){
 		return "open-tunnel";