+++ core/src/main/java/eu/unicore/ucc/authn/oidc/OIDCAgentProxy.java
@@ -18,35 +18,24 @@ import jnr.unixsocket.UnixSocketChannel;
 public class OIDCAgentProxy {
 
 	private static final String OIDC_SOCK = "OIDC_SOCK";
-	
+
 	public OIDCAgentProxy() {}
 
 	public static boolean isConnectorAvailable(){
-		return isConnectorAvailable(null);
+		return System.getenv(OIDC_SOCK)!=null;
 	}
 
-	public static boolean isConnectorAvailable(String usocketPath){
-		return System.getenv(OIDC_SOCK)!=null || usocketPath!=null;
-	}
-	
 	public String send(String data) throws Exception {
-		return send(data,System.getenv(OIDC_SOCK));
+		String path = System.getenv(OIDC_SOCK);
+		try(UnixSocketChannel channel = UnixSocketChannel.open(new UnixSocketAddress(path));
+				PrintWriter w = new PrintWriter(Channels.newOutputStream(channel));
+				InputStreamReader r = new InputStreamReader(Channels.newInputStream(channel)))
+		{
+			w.print(data);
+			w.flush();
+			CharBuffer result = CharBuffer.allocate(4096);
+			r.read(result);
+			return result.flip().toString();
+		}
 	}
-	
-	public String send(String data, String path) throws Exception {
-		UnixSocketAddress addr = new UnixSocketAddress(path);
-		UnixSocketChannel channel = UnixSocketChannel.create();
-		channel.connect(addr);
-        try(PrintWriter w = new PrintWriter(Channels.newOutputStream(channel));
-        	InputStreamReader r = new InputStreamReader(Channels.newInputStream(channel)))
-        {
-        	w.print(data);
-        	w.flush();
-        	CharBuffer result = CharBuffer.allocate(4096);
-        	r.read(result);
-        	result.flip();
-        	return result.toString();
-        }
-	}
-
 }