+++ core/src/test/java/eu/unicore/ucc/authn/TestAuthN.java
@@ -3,6 +3,7 @@ package eu.unicore.ucc.authn;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
 import java.util.Properties;
@@ -11,18 +12,33 @@ import org.apache.commons.io.FileUtils;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.junit.jupiter.api.Test;
 
+import eu.unicore.ucc.UCC;
 import eu.unicore.ucc.authn.oidc.TokenBasedAuthN;
-import eu.unicore.ucc.util.EmbeddedTestBase;
 
-public class TestAuthN extends EmbeddedTestBase {
+public class TestAuthN {
 
 	@Test
 	public void testNoAuthN() throws Exception {
+		assertTrue(UCC.getAuthNMethod(new NoAuthN().getName())!=null);
+		assertTrue(UCC.getAuthNMethod(new NoAuthN().getName()) instanceof NoAuthN);
+
 		var a = new NoAuthN();
 		a.setProperties(new Properties());
 		var m = new HttpGet("https://test");
 		a.addAuthenticationHeaders(m);
 		assertNull(m.getHeader("Authorization"));
+	}
+
+	@Test
+	public void testLoadAuthN() throws Exception {
+		assertTrue(UCC.getAuthNMethod(new TokenBasedAuthN().getName())!=null);
+		assertTrue(UCC.getAuthNMethod(new TokenBasedAuthN().getName()) instanceof TokenBasedAuthN);
+
+		assertTrue(UCC.getAuthNMethod(new KeystoreAuthN().getName())!=null);
+		assertTrue(UCC.getAuthNMethod(new KeystoreAuthN().getName()) instanceof KeystoreAuthN);
+
+		assertTrue(UCC.getAuthNMethod(new UsernameAuthN().getName())!=null);
+		assertTrue(UCC.getAuthNMethod(new UsernameAuthN().getName()) instanceof UsernameAuthN);
 	}
 
 	@Test